+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - moved SHOW_VIEW Event-Type from Coco.Event to Coco.RouterEvent
 - moved HIDE_VIEW Event-Type from Coco.Event to Coco.RouterEvent
 - added oldRoute object to Coco.RouterEvent
-- removed Coco.HashMap from library, use ES classes instead
+- removed Coco.HashMap from library, use ES6 classes instead
 
 ## 0.1.4